fix: policy server resource deleted before policies
There is an issue in the policy server controller which is causing
inconsistencies when the policy server is deleted. It's possible, due
the concurrent nature of events and controllers loops, where the
policies ended in a inconsistent state. Where the policy and its
webhooks are register in the cluster. But the policy server which should
run it is gone. Causing an issue when the control plane tries to contact
the policy server.

This is caused because when the policy server controller deletes all the
policy server policies, it filters out policies which have the deletion
timestamp field. Once no policies are found, in other words, when no
policies without deletion timestamp are found, the policy server
resources are deleted. However, the deletion timestamp does not ensure
that the resource is not register in the cluster. It tells that the
resource is marked to be deleted by the control plane. Which means that
we can reach a state where the policy server is deleted before the
policies. Causing issue when the control plane tries to send request to
the webhooks.This commit fixes that by removing the filter used to get
the policies in the deletion reconciliation loop.

+++ controllers/policyserver_controller.go
@@ -67,7 +67,7 @@ func (r *PolicyServerReconciler) Reconcile(ctx context.Context, req ctrl.Request
 		return ctrl.Result{}, fmt.Errorf("cannot retrieve policy server: %w", err)
 	}
 
-	policies, err := r.Reconciler.GetPolicies(ctx, &policyServer, admission.SkipDeleted)
+	policies, err := r.Reconciler.GetPolicies(ctx, &policyServer)
 	if err != nil {
 		return ctrl.Result{}, errors.Join(errors.New("could not get policies"), err)
 	}
@@ -100,29 +100,45 @@ func (r *PolicyServerReconciler) reconcile(ctx context.Context, policyServer *po
 }
 
 func (r *PolicyServerReconciler) reconcileDeletion(ctx context.Context, policyServer *policiesv1.PolicyServer, policies []policiesv1.Policy) (ctrl.Result, error) {
-	someDeletionFailed := false
+	if len(policies) != 0 {
+		// There are still policies scheduled on the PolicyServer, we have to
+		// wait for them to be completely removed before going further with the cleanup
+		return r.deletePoliciesAndRequeue(ctx, policyServer, policies)
+	}
+
+	if err := r.Reconciler.ReconcileDeletion(ctx, policyServer); err != nil {
+		return ctrl.Result{}, errors.Join(errors.New("could not reconcile policy server deletion"), err)
+	}
+
+	controllerutil.RemoveFinalizer(policyServer, constants.KubewardenFinalizer)
+	if err := r.Update(ctx, policyServer); err != nil {
+		// return if PolicyServer was previously deleted
+		if apierrors.IsConflict(err) {
+			return ctrl.Result{}, nil
+		}
+		return ctrl.Result{}, fmt.Errorf("cannot update policy server: %w", err)
+	}
+
+	return ctrl.Result{}, nil
+}
+
+func (r *PolicyServerReconciler) deletePoliciesAndRequeue(ctx context.Context, policyServer *policiesv1.PolicyServer, policies []policiesv1.Policy) (ctrl.Result, error) {
+	deleteError := make([]error, 0)
 	for _, policy := range policies {
+		if policy.GetDeletionTimestamp() != nil {
+			// the policy is already pending deletion
+			continue
+		}
 		if err := r.Delete(ctx, policy); err != nil && !apierrors.IsNotFound(err) {
-			someDeletionFailed = true
+			deleteError = append(deleteError, err)
 		}
 	}
-	if someDeletionFailed {
+
+	if len(deleteError) != 0 {
+		r.Log.Error(errors.Join(deleteError...), "could not remove all policies bound to policy server", "policy-server", policyServer.Name)
 		return ctrl.Result{}, fmt.Errorf("could not remove all policies bound to policy server %s", policyServer.Name)
 	}
-	if len(policies) == 0 {
-		if err := r.Reconciler.ReconcileDeletion(ctx, policyServer); err != nil {
-			return ctrl.Result{}, errors.Join(errors.New("could not reconcile policy server deletion"), err)
-		}
-		controllerutil.RemoveFinalizer(policyServer, constants.KubewardenFinalizer)
-		if err := r.Update(ctx, policyServer); err != nil {
-			// return if PolicyServer was previously deleted
-			if apierrors.IsConflict(err) {
-				return ctrl.Result{}, nil
-			}
-			return ctrl.Result{}, fmt.Errorf("cannot update policy server: %w", err)
-		}
-		return ctrl.Result{}, nil
-	}
+
 	return ctrl.Result{Requeue: true}, nil
 }
